ARUHA-759 Add per-event type authorization extension point
This extension point provides 2 methods: isAuthorized(..), which
determines whether a token is authorized to perform an operation
according to a list of subject attributes, at least one of which
the subject must hold to be authorized; and
isAuthorizationAttributeValid(..), which allows for the validation of
an authorization attribute. For example, one could check that the key
is a supported attribute type, and that the value refers to an entity
that actually exists (e.g., a username).

This extension point also provides a class for immutable
AuthorizationAttribute.