ARUHA-759 Improve authz plugin boundary
It is better for the plugin to provide a method to authorize a token
for an operation on an event type, rather than authorizing a token
against a list of attributes. This way, it is possible for the PDP to
make decisions that involve other factors than only subject attributes,
such as operation, resource, etc.