aruha-929 remove subject from the API
The subject is being removed from the API since we found its interface
to be limitating and not flexible enought to cover some use cases. Also
it's more convenient to use a subject available in the execution
context and not having to carry the subject all over the code.

The concept of a Subject, an entity representing the one that is
requesting access to a given resource, remains untouched. Only the
interface no longer includes it as a parameter.